Revert "Schedule the feed source update also at startup"
Turns out this can get triggered multiple times somehow which
then can result in concurrent execution of that task, something
the code isn't prepared for and thus causes duplicate entries.

As that's worse than no entries let's revert this while looking
for a better solution for the initial sources update.

This reverts commit e449d6f22706b41b425e1cdaa3ca23d7c35d3fe9.