Fix error cases in subscription heartbeat handling
We were previously also returning 200 for expired subscriptions due to
creating but not returning the error responses.

While at it, avoid the double query and instead check whether we managed
to update an entry to see whether we know a subscription.

Also, distinguish between invalid requests and the subscription not being
found in the response codes, and update the documentation to match that.