Add alert map view using pg_tileserv and MapLibre
At this point this merely shows the alert bounding boxes as they appear
in our database. This however already adds value as it shows problems
with bounding box wrapping at 180°.

The next step here is then to add more details about alerts to the
database (in particular the full geometry), which pg_tileserv will
automatically expose to the map.