Refactor CAP parsing code
As we'll need a much deeper view into the CAP content for the alert map,
move CAP parsing to actual CAP data types rather than dealing with raw XML
all over the place.

This isn't complete yet, we'd need the same for the geometry handling, but
that's a next step.