Color-code slow response times in the source list
Shows that we have a few feeds that take significantly longer than a
minute to update, which contributes to the Celery backlog problem. Also
shows that our network timeout handling doesn't catch all cases yet
apparently.

Also, show the total sum of the fetch duration. With 4 workers this must
not exceed 240s (roughly 1sec per feed), we are significantly above that
currently.