Measure the entire time we need for processing a feed
As ultimately that is what decides whether we are fast enough in
processing alerts or not.

Especially, also record the measurement in case of errors. We otherwise
end up with rather misleading data for feeds that time out.

Shows that our timeout handling actually works reliably, it's just the
measurements being wrong.