Don't download alerts we already know
This works for CAP feeds that include the CAP identifier and sent time
in the feed already, which is the case for several Meteoalarm feeds for
example.

The impact is particularly big on unchanged feeds with very expensive to
process geometry such as Latvia, going from minutes down to less than a
second.