Add geometry simplification infrastructure
Not applied yet to our checked in data, but already useful to validate
the approach and determine the correct parameters. Even with the very
defensive defaults it's achieving 10-150x reductions on the particularly
large coastal geometries, with the resulting polygon still fully containing
the original one.

The basic idea is the following 4 steps:
(1) Expand the polygon. This is very effective in removing bays/fjords/etc
    and "swallowing" islands.
(2) Apply the Douglas-Peucker simplification. This reduces the level of
    detail of the polygon lines. Using a smaller threshold than use for
    expansion guarantees we still contain the original geometry.
(3) Shrink the result again. This is to get back closer to the original
    geometry and should be done by the amount we expanded minus half the
    simplification threshold to still cover the original geometry.
(4) Round coordinate decimals to an appropriate length. This is computed
    to be enough to represent 1/10th of the simplification threshold.
    That's still very defensive but at the same time reasonably effective.

While this is currently mainly used via its own command line tool, most of
the logic is in library code and could be directly added to the data update
scripts if needed.